modify : user 패키지명을 auth 로 변경
사용자 비즈니스 로직과 로그인, 인증 관련 코드를 분리하기 위함
