+++ src/main/java/com/yanolja/areas/auth/dto/TokenRefreshRequest.java
@@ -1,4 +1,4 @@
-package com.yanolja.areas.user.dto;
+package com.yanolja.areas.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LogoutRequest {
+public class TokenRefreshRequest {
     
     @NotBlank(message = "리프레시 토큰은 필수입니다.")
     private String refreshToken;
