+++ src/main/java/com/yanolja/areas/auth/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.yanolja.areas.user.repository;
+package com.yanolja.areas.auth.repository;
 
-import com.yanolja.areas.user.domain.User;
+import com.yanolja.areas.auth.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
